--- src/bach/index.js
@@ -1,11 +1,0 @@
-const { series } = require('./lib/series');
-const { parallel } = require('./lib/parallel');
-const { settleSeries } = require('./lib/settleSeries');
-const { settleParallel } = require('./lib/settleParallel');
-
-module.exports = {
-  series,
-  parallel,
-  settleSeries,
-  settleParallel,
-};
