+++ src/bach/series.js
@@ -1,30 +1,26 @@
-const { runFunction } = require('../../run-function');
+const { runFunction } = require('../run-function');
 const { parseOptions } = require('./parseArgs');
 
-function parallel(funcs, options) {
+function series(funcs, options) {
   const normalizeOptions = parseOptions(options);
   return (done) => {
     async function run() {
       const results = new Array(funcs.length).fill(undefined);
       let error = null;
       try {
-        await Promise.all(
-          funcs.map(async (fn, idx) => {
-            results[idx] = await runFunction(fn, idx, normalizeOptions);
-          }),
-        );
+        for (let idx = 0; idx < funcs.length; idx++) {
+          results[idx] = await runFunction(funcs[idx], idx, normalizeOptions);
+        }
       } catch (e) {
         error = e;
       }
       return [error, results];
     }
 
-    run().then(([error, results]) => {
-      return done(error, results);
-    });
+    run().then(([error, results]) => done(error, results));
   };
 }
 
 module.exports = {
-  parallel,
+  series,
 };
